Move back OnHandPickUp to server-only in order to accommodate some form of mispredict? Not sure what effect is talked about.
