+++ Content.Client/_Starlight/Lube/LubedSystem.cs
@@ -1,0 +1,5 @@
+﻿using Content.Shared._Starlight.Lube;
+
+namespace Content.Client._Starlight.Lube;
+
+public sealed partial class LubedSystem : SharedLubedSystem;
